\name{basefun-package}
\docType{package}
\alias{basefun-package}
\alias{basefun}
\title{General Information on the \pkg{basefun} Package}
\description{
  The \pkg{basefun} package offers a small collection of objects for
  handling basis functions and corresponding methods.
}
\details{
  The package was written to support the \pkg{mlt}
  package \bibcitep{basefun::Hothorn:2018}
  and will be of limited use outside this package.
}
\references{\bibshow{*}}
\keyword{package}
