\name{basefun-package}
\docType{package}
\alias{basefun-package}
\alias{basefun}
\title{General Information on the \pkg{basefun} Package}
\description{
  The \pkg{basefun} package offers a small collection of objects for
  handling basis functions and corresponding methods.

  The package was written to support the \pkg{mlt}
  package and will be of limited use outside this package.
  The user interface is not yet stable, so please be careful using 
  \pkg{basefun} in your package.
}
\author{
  This package is authored by Torsten Hothorn <Torsten.Hothorn@R-project.org>.
}
\keyword{package}
